Add a new demo similar to Bubbles!!!
In the spirit of
http://ie.microsoft.com/testdrive/Performance/Bubbles/Default.html
this demo animates a number of sprites bouncing off the walls and each
other. When a pair of bubbles collide, part of the background is
revealed.

The original Microsoft demo computes the blur on the CPU (presumably to
highlight a weakness in chrome and firefox), this demo instead performs
all pixel manipulation on the GPU.